19- change nfbasis(T, flag, fa) to nfbasis(T, listP).
flag was used to invoke round2 instead of round4 (inefficient=> useless)
OR to only partially factor poldisc(T), up to primelimit (very dangerous
since primelimit is a global variable). Now listP describes a list of
primes, and we return a basis of an order which is p-maximal at all
those primes: either a vector of primes, a factorisation (as fa before)
or an integer B to indicated {p <= B} (a safe and flexible version of
nfbasis(T, 1)). nfdisc() was changed similarly.