Test coverage reporting [#1470] by Esa Peuha
Quoting Esa:
Here is a patch that uses gcov and lcov for test coverage reporting; gcov
comes with gcc and does most of the work, while lcov is a separate program
that does the html formatting. Using them is simple: in the main pari
directory, execute

./Configure -gcov
make gcov

then in Oxxx.gcov

make test-all
make lcov-report

after which lcov-report/index.html is the starting point to the report.